FSAL: Use Filesystem_Id in NFS file handles
If Filesystem_id is specified in an EXPORT, use it for
the filesystem fsid.  This allows multiple directories
to be exported from the same underlying filesystem.

Add config option nfs_param.core_param.fsid_override to
enable the above behaviour.

I have had this patch for a while.  I'll submit it
for consideration.

Change-Id: I8505600a77e56fd4cb8214b251bad3a782a98695
Signed-off-by: Peter Schwenke <pschwenke@ddn.com>